% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_inflation.R
\name{adjust_inflation}
\alias{adjust_inflation}
\title{Adjust Values for Inflation}
\usage{
adjust_inflation(amounts, inflation_rate)
}
\arguments{
\item{amounts}{Numeric vector. The list of future money values you want to adjust}

\item{inflation_rate}{Numeric. The annual inflation rate (as a decimal).}
}
\value{
A list of values showing what that money is worth in today's economy.
}
\description{
Converts future dollar amounts into "today's dollars" to show their
actual purchasing power.
}
