% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_wealth.R
\name{calc_wealth}
\alias{calc_wealth}
\title{Calculate Wealth Trajectory}
\usage{
calc_wealth(principal, monthly, rate, years)
}
\arguments{
\item{principal}{Numeric. The starting balance of the account.}

\item{monthly}{Numeric. The amount contributed at the end of each month.}

\item{rate}{Numeric. The annual nominal interest rate (as a decimal, e.g., 0.05 for 5 parcent).}

\item{years}{Numeric. The number of years to project.}
}
\value{
A numeric vector representing the account balance at each month.
}
\description{
Calculate Wealth Trajectory
}
