% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_savings.R
\name{project_savings}
\alias{project_savings}
\title{Project savings with regular contributions}
\usage{
project_savings(start_balance, monthly_contribution, annual_rate, years)
}
\arguments{
\item{start_balance}{Initial balance.}

\item{monthly_contribution}{Monthly contribution.}

\item{annual_rate}{Annual return rate as decimal.}

\item{years}{Number of years.}
}
\value{
Data frame with monthly savings path.
}
\description{
Project savings with regular contributions
}
\examples{
project_savings(5000, 250, 0.06, 20)
}
