% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_forecast.R
\name{summarize_forecast}
\alias{summarize_forecast}
\title{Summarize a forecast}
\usage{
summarize_forecast(forecast, initial)
}
\arguments{
\item{forecast}{A data frame from grow_investment or project_savings.}

\item{initial}{Starting value.}
}
\value{
One-row summary data frame.
}
\description{
Summarize a forecast
}
\examples{
x <- grow_investment(1000, 0.07, 5, 200)
summarize_forecast(x, 1000)
}
