% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_wealth.R
\name{calc_wealth}
\alias{calc_wealth}
\title{Calculate Wealth Trajectory}
\usage{
calc_wealth(principal, monthly, rate, years)
}
\arguments{
\item{principal}{Numeric. Your starting balance (how much money you have now)}

\item{monthly}{Numeric. The amount you plan to save/add every month.}

\item{rate}{Numeric. The annual nominal interest rate (as a decimal, e.g., 0.05 for 5 parcent).}

\item{years}{Numeric. How many years into the future you want to project.}
}
\value{
A list of monthly balances over the specified time.
}
\description{
Calculate Wealth Trajectory
}
